+++ src/allegrord.cpp
@@ -32,7 +32,7 @@ public:
     bool parse();
     long parse_chan(string &field);
     long parse_int(string &field);
-    int find_real_in(string &field, int n);
+    size_t find_real_in(string &field, size_t n);
     double parse_real(string &field);
     void parse_error(string &field, long offset, const char *message);
     double parse_dur(string &field, double base);
@@ -43,7 +43,7 @@ public:
     long parse_after_key(int key, string &field, int n);
     long find_int_in(string &field, int n);
     bool parse_attribute(string &field, Alg_parameter *parm);
-    bool parse_val(Alg_parameter *param, string &s, int i);
+    bool parse_val(Alg_parameter *param, string &s, size_t i);
     bool check_type(char type_char, Alg_parameter *param);
 };
 
@@ -472,13 +472,13 @@ long Alg_reader::parse_int(string &field)
 }
 
 
-int Alg_reader::find_real_in(string &field, int n)
+size_t Alg_reader::find_real_in(string &field, size_t n)
 {
     // scans from offset n to the end of a real constant
     bool decimal = false;
-    int len = field.length();
+    size_t len = field.length();
     if (n < len && field[n] == '-') { n++; } // parse one minus sign
-    for (int i = n; i < len; i++) {
+    for (size_t i = n; i < len; i++) {
         char c = field[i];
         if (!isdigit(c)) {
             if (c == '.' && !decimal) {
@@ -686,9 +686,9 @@ bool Alg_reader::parse_attribute(string &field, Alg_parameter *param)
 }
 
 
-bool Alg_reader::parse_val(Alg_parameter *param, string &s, int i)
+bool Alg_reader::parse_val(Alg_parameter *param, string &s, size_t i)
 {
-    int len = static_cast<int>(s.length());
+    size_t len = s.length();
     if (i >= len) {
         return false;
     }
@@ -719,7 +719,7 @@ bool Alg_reader::parse_val(Alg_parameter *param, string &s, int i)
             return false;
         }
     } else if (isdigit(s[i]) || s[i] == '-' || s[i] == '.') {
-        int pos = i;
+        size_t pos = i;
         bool period = false;
         if (s[pos] == '-') { pos++; }
         while (pos < len) {
